+++ docs/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 ---------
 
+* 0.0.2
+
+  - Make testing layer actually return the headers and bodies that
+    are programmatically added via ``glinda.testing.services.Response``.
+
 * 0.0.1
 
   - Add :class:`glinda.testing.services.Service`